@StephanDollberg 2023-11-15
![image](https://github.com/redpanda-data/seastar/assets/2904723/781aa41d-f554-4e47-b8fa-468e538daf1b)
![image](https://github.com/redpanda-data/seastar/assets/2904723/e3fb68fa-73ed-4e9e-90a2-b2591a5cf3e8)



Some metrics - first three runs are dev and last three with this patch (ignore stuff in the middle)

![image](https://github.com/redpanda-data/seastar/assets/2904723/1ecb729c-1729-4775-aeb1-4f2773e603b9)
![image](https://github.com/redpanda-data/seastar/assets/2904723/3d37efa3-756f-471f-a0b6-718b89bf53ec)
![image](https://github.com/redpanda-data/seastar/assets/2904723/ced5e6d6-ced1-4c6a-a4fe-534bf5f0c189)