@travisdowns 2023-11-17
Great analysis, this chart in particular is amazing:

![image](https://github.com/redpanda-data/seastar/assets/2403521/19da39c6-7432-4aea-9ad0-696c3dde20ef)

The impact in p99 latency is close to 3x for this benchmark: it really shows that you can't just look at reactor utilization and extrapolate to latency impacts.